changed import order and install script

+++ install.sh
@@ -1,23 +1,5 @@
 #!/bin/bash
 
-# Run Uninstall Script
-# Find all .files
-# If original file exists, back it up to {file}.dtbak
-# Symlink new .file
-
-# ./uninstall.sh
-# ./install_packages.sh
-
-# for file in $(find . -maxdepth 1 -name ".*" -type f  -printf "%f\n" ); do
-#     if [ -e ~/$file ]; then
-#         mv -f ~/$file{,.dtbak}
-#     fi
-#     ln -s $PWD/$file ~/$file
-# done
-
-
-# echo "Backed up original files and created symlinks."
-
 # if not installed, install tmux, zshell
 if command -v zsh > /dev/null; then
     echo "zsh already installed"
@@ -30,6 +12,28 @@ if command -v tmux > /dev/null; then
 else
     echo "installing tmux"
     sudo apt install tmux -y
+fi
+
+# build neovim from source if not installed
+if command -v nvim > /dev/null; then
+    echo "neovim already installed"
+else
+    sudo apt-get install ninja-build gettext libtool libtool-bin autoconf automake cmake g++ pkg-config unzip curl doxygen
+    mkdir repos
+    cd repos
+    if [ ! -d neovim ]; then
+        git clone https://github.com/neovim/neovim
+    fi
+    cd neovim
+    make CMAKE_BUILD_TYPE=RelWithDebInfo
+    sudo make install
+    cd ../..
+fi
+
+# symlink nvim
+if [ ! -L ~/.config/nvim ]; then
+    echo "Create nvim symlink"
+    sudo ln -s $PWD/nvim ~/.config/nvim
 fi
 
 # symlink oh-my-posh
@@ -46,7 +50,9 @@ fi
 
 # symlink other dotfiles
 for file in $(find . -maxdepth 1 -name ".*" -type f  -printf "%f\n" ); do
-    ln -s $PWD/$file ~/$file
+    if [ ! -L ~/$file ] && [ ! $file != ".gitignore" ]; then
+        ln -s $PWD/$file ~/$file
+    fi
 done
 
 # set zsh as default shell
